inmobiliaria-add_DRF(3)/inmobiliaria-add_DRF/inmobiliaria-add_DRF/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
djangorestframework==3.15.0
install==1.3.5
kombu==5.3.5
pillow==10.3.0
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
